Разделяет тесты для редьюсера
